+++ src/behavior/default.rs
@@ -20,7 +20,7 @@ impl Behavior for DefaultBehavior {
 		let our_agent = world_view.our_agent().clone();
 
 		// more likely to go right
-		let random_angle = Radians::from(Deg(10.0)) * (thread_rng().gen_range(-1i8..=2) as f64);
+		let random_angle = Radians::from(Deg(10.0)) * (thread_rng().gen_range(-1i8..=2) as f32);
 
 		if world_view.our_id() != world_view.current_it() {
 			// we're not "it", run in a random direction with full speed
